% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_query.R
\name{qb_query}
\alias{qb_query}
\title{Query quickbase tables}
\usage{
qb_query(
  subdomain,
  user_token,
  table_id,
  query = NULL,
  field_ids = NULL,
  agent = NULL
)
}
\arguments{
\item{subdomain}{A character vector with one element. Found at the beginning of the Quickbase URL. Realm specific.}

\item{user_token}{A character vector with one element. Created in 'My Preferences' under 'Manage user tokens' link.}

\item{table_id}{A character vector of table ids. Found in the 'Reports & Charts' page in Quickbase and in the report URL.}

\item{query}{Optional. A character vector with one element. A query contructed using Quickbase's query language.}

\item{field_ids}{Optional. A vector of integers representing the field ids you want returned.}

\item{agent}{Optional. A character vector with one element.}
}
\value{
A list of tibbles with queried data
}
\description{
Query quickbase tables
}
