% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_run.R
\name{qb_run}
\alias{qb_run}
\title{Run a Quickbase report}
\usage{
qb_run(
  subdomain,
  user_token,
  table_id,
  report_id,
  agent = NULL,
  skip = NULL,
  top = NULL,
  json_out = FALSE,
  flat = TRUE,
  type_suffix = FALSE
)
}
\arguments{
\item{subdomain}{Character vector with one element. Found at the beginning of
the Quickbase URL. Realm specific.}

\item{user_token}{Character vector with one element. Created in 'My
Preferences' under 'Manage user tokens' link.}

\item{table_id}{Character vector with one element. Found in the URL of a
Quickbase table between /db/ and ?}

\item{report_id}{Character vector with one element. Found in the 'Reports &
Charts' page in Quickbase and in the report URL.}

\item{agent}{Optional. Character vector with one element. Describes
user/agent making API call.}

\item{skip}{Optional. Integer. The number of rows to skip from the top of a
record set.}

\item{top}{Optional. Integer. The limit on the number of records to pull
starting at the top of a record set.}

\item{json_out}{Optional. Logical. Set TRUE if you want JSON returned instead
of a tibble}

\item{flat}{Optional. Logical. Set TRUE to reduce nesting in the payload.}

\item{type_suffix}{Optional. Logical. Set TRUE to append each field label
with its Quickbase data type.}
}
\value{
A tibble or JSON object.
}
\description{
\code{qb_run} tells the Quickbase API to run a report and returns its data.
}
\section{Warning}{
 Extracting nested fields (e.g., user-lists) will result a
nested tibble.
}

\examples{
# Minimum required info
my_tibble <- qb_run(subdomain = "bhi",
    user_token = keyring::key_get("qb_example"),
    table_id = "bn9d8iesz",
    report_id = "1")

# Long form call
my_tibble <- qb_run(subdomain = "bhi.quickbase.com",
    user_token = paste0("QB-USER-TOKEN ", keyring::key_get("qb_example")),
    table_id = "bn9d8iesz",
    report_id = "1",
    agent = "FileService_Integration_V2.1",
    skip = 2,
    top = 10,
    json_out = FALSE,
    flat = TRUE)
}
\references{
\href{https://developer.quickbase.com/}{Quickbase API
documentation}
}
