% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbr-package.R
\docType{package}
\name{qbr-package}
\alias{qbr}
\alias{qbr-package}
\title{qbr: Access the 'Quickbase' JSON API}
\description{
Programmatically access the 'Quickbase' JSON API \url{https://developer.quickbase.com}. You supply parameters for an API call, 'qbr' delivers an http request to the API endpoint and returns its response. Outputs follow 'tidyverse' philosophy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BHII-KSC/qbr}
  \item Report bugs at \url{https://github.com/BHII-KSC/qbr/issues}
}

}
\author{
\strong{Maintainer}: John Erdmann \email{john.erdmann@keene.edu}

Other contributors:
\itemize{
  \item Keene State College [copyright holder, funder]
}

}
\keyword{internal}
