% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbr.R
\docType{package}
\name{qbr}
\alias{qbr}
\title{qbr: A package for interacting with the 'Quickbase' JSON API}
\description{
The qbr package is currently focused on extracting data from 'Quickbase' via
its JSON API.
}
\section{qbr functions}{

* \code{qb_run} Runs a 'Quickbase' report and returns its data as a
tibble.
}

