% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_run.R
\name{qb_run}
\alias{qb_run}
\title{Run a Quickbase report}
\usage{
qb_run(
  table_id,
  report_id,
  user_token,
  subdomain,
  skip = NULL,
  top = NULL,
  agent = NULL
)
}
\arguments{
\item{table_id}{A character vector with one element. Found in the URL of a Quickbase table between /db/ and ?}

\item{report_id}{A character vector with one element. Found in the 'Reports & Charts' page in Quickbase and in the report URL.}

\item{user_token}{A character vector with one element. Created in 'My Preferences' under 'Manage user tokens' link.}

\item{subdomain}{A character vector with one element. Found at the beginning of the Quickbase URL. Realm specific.}

\item{skip}{Optional. An integer. The number of rows to skip from the top of a record set.}

\item{top}{Optional. An integer. The limit on the number of records to pull starting top of a record set.}

\item{agent}{Optional. A character vector with one element.}
}
\value{
A tibble with the report's data.
}
\description{
Run a Quickbase report
}
\examples{
# Minimum required info
my_tibble <- qb_run(table_id = "bn9d8iesz",
    report_id = "1",
    user_token = "b25vav_itkb_wux8ipccd4345dsaq7dvczqvddz",
    subdomain = "bhi.quickbase.com")

# Full argument set
my_tibble <- qb_run(table_id = "bn9d8iesz",
    report_id = "1",
    user_token = "b25vav_itkb_wux8ipccd4345dsaq7dvczqvddz",
    subdomain = "bhi",
    skip = 2,
    top = 10,
    agent = "FileService_Integration_V2.1")
}
