% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbr.R
\docType{package}
\name{qbr}
\alias{qbr}
\title{qbr: A package for interacting with the Quickbase API}
\description{
The qbr package is currently focused on extracting data from Quickbase via the API.
}
\section{qbr functions}{

* \code{qb_run} Runs a quickbase report and returns its data as a
tibble or JSON.
* \code{qb_query} Submits a user-defined query to the Quickbase API
and returns the result.
}

